% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.R
\name{fit_mgsm}
\alias{fit_mgsm}
\title{Fits a mixed generalized survival model with the probit link}
\usage{
fit_mgsm(
  formula,
  data,
  id,
  rng_formula,
  df = 4L,
  maxpts = c(1000L, 10000L),
  key = 3L,
  abseps = 0,
  releps = 0.001,
  seed = 1L
)
}
\arguments{
\item{formula}{formula for the fixed effects. Must contain the intercept.}

\item{data}{a data.frame with the needed variables.}

\item{id}{the cluster identifier for each individual.}

\item{rng_formula}{formula for the design matrix for the random effects.}

\item{df}{the degrees of freedom to use for the spline in the generalized
survival model.}

\item{maxpts}{a vector of integer with successive number of samples to use
when fitting the model.}

\item{key, abseps, releps}{parameters for the Monte Carlo method.}

\item{seed}{the fixed seed to use.}
}
\description{
Fits a mixed generalized survival model with the probit link
}
